factorize the call to the identity service when creating an admin client
